Create for a new unique key
Changed Database engine to allow for a new unique key 'device_user_pc'. The new key supports a related change in the API (index.php) so that we can use INSERT ON DUPLICATE KEY UPDATE to eliminate multiple entries for the same USB on the same computer for the same person. In this case we are now updating the "original" record...